SystemUI: Enable SystemUI
Enabled SystemUI
Fixed SystemUI compilation

Conflicts:
	packages/SystemUI/res/layout/quick_settings_brightness_dialog.xml
	packages/SystemUI/res/layout/recents_task_view.xml
	packages/SystemUI/res/layout/signal_cluster_view.xml
	packages/SystemUI/res/layout/zen_mode_panel.xml
	packages/SystemUI/src/com/android/systemui/volume/VolumePanel.java

+++ packages/SystemUI/src/com/android/systemui/qs/tiles/DataTile.java
@@ -84,7 +84,7 @@ public class DataTile extends QSTile<QSTile.BooleanState> {
         for (int i = 0; i < phoneCount; i++) {
             Settings.Global.putInt(mContext.getContentResolver(),
                     Settings.Global.MOBILE_DATA + i, (enabled) ? 1 : 0);
-            long[] subId = SubscriptionManager.getSubId(i);
+            int[] subId = SubscriptionManager.getSubId(i);
             mTelephonyManager.setDataEnabledUsingSubId(subId[0], enabled);
         }
     }
